dokoncene

+++ Matus_Remen/FilterUJVekovaSkupinaPodiel.h
@@ -1,4 +1,4 @@
-/*
+
 #pragma once
 #include "KriteriumUjVekovaSkupinaPodiel.h"
 #include "VZDELANIE_enum.h"
@@ -7,17 +7,17 @@
 
 
 
-class FilterUJVekovaSkupinaPodiel : public FilterInterval<UzemnaJednotka, int> {
+class FilterUJVekovaSkupinaPodiel : public FilterInterval<UzemnaJednotka, double> {
 
 public:
 
-	FilterUJVekovaSkupinaPodiel(EVS_ENUM vekovaSkupina, int minimum, int maximum);
+	FilterUJVekovaSkupinaPodiel(EVS_ENUM vekovaSkupina, double minimum, double maximum);
 
 };
 
- FilterUJVekovaSkupinaPodiel::FilterUJVekovaSkupinaPodiel(EVS_ENUM vekovaSkupina, int minimum, int maximum) :
+ FilterUJVekovaSkupinaPodiel::FilterUJVekovaSkupinaPodiel(EVS_ENUM vekovaSkupina, double minimum, double maximum) :
 	 FilterInterval(new KriteriumUJVekovaSkupinaPodiel(vekovaSkupina), minimum, maximum)
 {
 }
 
- */
+
